Move error accumulation to g_input
