+++ src/g_input.c
@@ -16,6 +16,7 @@
 //
 
 #include <math.h>
+#include <string.h>
 
 #include "d_main.h"
 #include "g_game.h"
@@ -24,6 +25,7 @@
 #include "i_video.h"
 #include "m_config.h"
 #include "r_main.h"
+#include "r_state.h"
 
 //
 // Side Movement
@@ -53,6 +55,114 @@ void G_UpdateSideMove(void)
         G_RoundSide = RoundSide_Full;
         sidemove = autostrafe50 ? forwardmove : default_sidemove;
     }
+}
+
+//
+// Error Accumulation
+//
+
+typedef struct
+{
+    double angle;
+    double pitch;
+    double side;
+    double vert;
+} carry_t;
+
+static carry_t prevcarry, carry;
+
+void G_UpdateCarry(void)
+{
+    prevcarry = carry;
+}
+
+void G_ClearCarry(void)
+{
+    memset(&prevcarry, 0, sizeof(prevcarry));
+    memset(&carry, 0, sizeof(carry));
+}
+
+static int CarryError(double value, const double *prevcarry, double *carry)
+{
+    const double desired = value + *prevcarry;
+    const int actual = lround(desired);
+    *carry = desired - actual;
+    return actual;
+}
+
+static short CarryAngleTic_Full(double angle)
+{
+    return CarryError(angle, &prevcarry.angle, &carry.angle);
+}
+
+static short CarryAngle_Full(double angle)
+{
+    const short fullres = CarryAngleTic_Full(angle);
+    localview.angle = fullres << FRACBITS;
+    return fullres;
+}
+
+// Round to nearest 256 for single byte turning. From Chocolate Doom.
+#define BYTE_TURN(x) (((x) + 128) & 0xFF00)
+
+static short CarryAngle_FakeLongTics(double angle)
+{
+    return (localview.angleoffset = BYTE_TURN(CarryAngle_Full(angle)));
+}
+
+static short CarryAngleTic_LowRes(double angle)
+{
+    const double fullres = angle + prevcarry.angle;
+    const short lowres = BYTE_TURN((short)lround(fullres));
+    carry.angle = fullres - lowres;
+    return lowres;
+}
+
+static short CarryAngle_LowRes(double angle)
+{
+    const short lowres = CarryAngleTic_LowRes(angle);
+    localview.angle = lowres << FRACBITS;
+    return lowres;
+}
+
+short (*G_CarryAngleTic)(double angle);
+short (*G_CarryAngle)(double angle);
+
+void G_UpdateAngleFunctions(void)
+{
+    G_CarryAngleTic = lowres_turn ? CarryAngleTic_LowRes : CarryAngleTic_Full;
+    G_CarryAngle = G_CarryAngleTic;
+
+    if (!netgame || solonet)
+    {
+        if (lowres_turn && fake_longtics)
+        {
+            G_CarryAngle = CarryAngle_FakeLongTics;
+        }
+        else if (uncapped && raw_input)
+        {
+            G_CarryAngle = lowres_turn ? CarryAngle_LowRes : CarryAngle_Full;
+        }
+    }
+}
+
+int G_CarryPitch(double pitch)
+{
+    return (localview.pitch =
+                CarryError(pitch, &prevcarry.pitch, &carry.pitch));
+}
+
+int G_CarrySide(double side)
+{
+    const double desired = side + prevcarry.side;
+    const int actual = G_RoundSide(desired);
+    carry.side = desired - actual;
+    return actual;
+}
+
+int G_CarryVert(double vert)
+{
+    return CarryError(vert, &prevcarry.vert, &carry.vert);
 }
 
 //
